Allocate during initialization.
We were seeing some errors during MacOS CI in release mode
using the memory_tools.  After doing some local debugging, I found
that it was segfaulting while trying to access a pthread variable
during initialization.  Looking at the code, it looks like it
is possible that the constructor in libmemory_interpose gets
called *before* the static constructor for the recursive_mutex
in unix_common.cpp, leading to this problem.  To work around
this, introduce a new function called complete_static_initialization()
which allocates the recursive_mutex and then sets
initialization_complete to true.  This should be safe from races between
construction and malloc/realloc/calloc on other threads, since those
won't be able to access the recursive_mutex until
initialization_complete is set to true.  It is not safe from races from
multiple copies of __apple_memory_tools_init/__linux_memory_tools_init,
but the rest of the code also assumes that and we don't think it is
possible.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>